account: Update error handling and tests.
This commit updates error handling where the requests to the
CloudAPI are being made, especially around handling the closure
of the underlying `io.ReadCloser` when the body is returned
as part of the response.

Additionally, updates tests where needed so that they are
consistent, plus compliant with the behaviour of the HTTP
client from Go's standard library (e.g., handling of
`Response.Body`, etc.).

Signed-off-by: Krzysztof Wilczynski <kw@linux.com>